requirements.txt
coloredlogs==10.0
requests==2.18.4
mysql-connector-python==8.0.11
